requirements.txt
discord.py=2.3.2
loguru=0.7.2
python-dotenv=1.0.1